+++ packages/babel-plugin-lingui-macro/README.md
@@ -1,0 +1,40 @@
+[![License][badge-license]][license]
+[![Version][badge-version]][package]
+[![Downloads][badge-downloads]][package]
+
+# @lingui/babel-plugin-lingui-macro
+
+> Babel plugin that does actual transforms of Lingui's Macros
+
+`@lingui/babel-plugin-lingui-macro` is part of [LinguiJS][linguijs]. See the [documentation][documentation] for all information, tutorials and examples.
+
+## Installation
+
+```sh
+npm install --save-dev @lingui/babel-plugin-lingui-macro
+# yarn add --dev @lingui/babel-plugin-lingui-macro
+```
+
+## Usage
+
+### Via `.babelrc`
+
+**.babelrc**
+
+```json
+{
+  "plugins": ["@lingui/babel-plugin-lingui-macro"]
+}
+```
+
+## License
+
+[MIT][license]
+
+[license]: https://github.com/lingui/js-lingui/blob/main/LICENSE
+[linguijs]: https://github.com/lingui/js-lingui
+[documentation]: https://lingui.dev
+[package]: https://www.npmjs.com/package/@lingui/babel-plugin-lingui-macro
+[badge-downloads]: https://img.shields.io/npm/dw/@lingui/babel-plugin-lingui-macro.svg
+[badge-version]: https://img.shields.io/npm/v/@lingui/babel-plugin-lingui-macro.svg
+[badge-license]: https://img.shields.io/npm/l/@lingui/babel-plugin-lingui-macro.svg
