+++ packages/babel-plugin-lingui-macro/src/constants.ts
