+++ packages/babel-plugin-lingui-macro/src/icu.test.ts
