+++ packages/babel-plugin-lingui-macro/src/icu.ts
