+++ packages/babel-plugin-lingui-macro/src/index.ts
