+++ packages/babel-plugin-lingui-macro/src/macroJs.test.ts
