+++ packages/babel-plugin-lingui-macro/src/macroJs.ts
