+++ packages/babel-plugin-lingui-macro/src/macroJsx.ts
