+++ packages/babel-plugin-lingui-macro/src/utils.ts
