+++ packages/babel-plugin-lingui-macro/test/fixtures/js-t-continuation-character.expected.js
