+++ packages/babel-plugin-lingui-macro/test/fixtures/js-t-var/js-t-var.expected.js
