+++ packages/babel-plugin-lingui-macro/test/fixtures/jsx-plural-select-nested.js
