+++ packages/babel-plugin-lingui-macro/test/js-defineMessage.ts
