+++ packages/babel-plugin-lingui-macro/test/js-plural.ts
