+++ packages/babel-plugin-lingui-macro/test/js-select.ts
