+++ packages/babel-plugin-lingui-macro/test/js-selectOrdinal.ts
