+++ packages/babel-plugin-lingui-macro/test/js-t.ts
