+++ packages/babel-plugin-lingui-macro/test/js-useLingui.ts
