+++ packages/babel-plugin-lingui-macro/test/jsx-plural.ts
