+++ packages/babel-plugin-lingui-macro/test/jsx-select.ts
