+++ packages/babel-plugin-lingui-macro/test/jsx-selectOrdinal.ts
