+++ packages/babel-plugin-lingui-macro/test/jsx-trans.ts
