+++ packages/babel-plugin-lingui-macro/test/lingui.config.js
