+++ packages/macro/src/index.test.ts
@@ -1,0 +1,7 @@
+import macro from "@lingui/macro"
+
+describe("react-macro", () => {
+  it("Should re-export Macro", () => {
+    expect((macro as any).isBabelMacro).toBeTruthy()
+  })
+})
