+++ packages/react-macro/README.md
@@ -38,8 +38,8 @@ function MyComponent() {
 [linguijs]: https://github.com/lingui/js-lingui
 [documentation]: https://lingui.dev
 [reference]: https://lingui.dev/ref/macro/
-[package]: https://www.npmjs.com/package/@lingui/macro
-[badge-downloads]: https://img.shields.io/npm/dw/@lingui/macro.svg
-[badge-version]: https://img.shields.io/npm/v/@lingui/macro.svg
-[badge-license]: https://img.shields.io/npm/l/@lingui/macro.svg
+[package]: https://www.npmjs.com/package/@lingui/react-macro
+[badge-downloads]: https://img.shields.io/npm/dw/@lingui/react-macro.svg
+[badge-version]: https://img.shields.io/npm/v/@lingui/react-macro.svg
+[badge-license]: https://img.shields.io/npm/l/@lingui/react-macro.svg
 [badge-macro]: https://img.shields.io/badge/babel--macro-%F0%9F%8E%A3-f5da55.svg
