+++ tsconfig.json
@@ -20,7 +20,8 @@
       "@lingui/cli/api": ["./packages/cli/src/api"],
       "@lingui/react": ["./packages/react/src"],
       "@lingui/conf": ["./packages/conf/src"],
-      "@lingui/macro/plugin": ["./packages/macro/src/plugin.ts"],
+      "@lingui/babel-plugin-lingui-macro": ["./packages/babel-plugin-lingui-macro/src/index.ts"],
+      "@lingui/babel-plugin-lingui-macro/macro": ["./packages/babel-plugin-lingui-macro/src/macro.ts"],
       "@lingui/macro": ["./packages/macro/src"],
       "@lingui/react-macro": ["./packages/react-macro/src"],
       "@lingui/format-po": ["./packages/format-po/src/po.ts"],
